colplus-admin/src/test/java/org/col/admin/task/importer/acef/AcefReaderTest.java
package org.col.admin.task.importer.acef;

import com.google.common.collect.Lists;
import org.col.api.model.TermRecord;
import org.gbif.dwc.terms.AcefTerm;
import org.gbif.utils.file.FileUtils;
import org.junit.Ignore;
import org.junit.Test;

import java.util.List;
import java.util.Optional;

import static org.junit.Assert.*;

/**
 *
 */
public class AcefReaderTest {

  @Test
  public void fromFolder() throws Exception {
    AcefReader reader = AcefReader.from(FileUtils.getClasspathFile("acef/0"));

    int counter = 0;
    for (TermRecord tr : reader.read(AcefTerm.AcceptedSpecies)){
      counter++;
      assertNotNull(tr.get(AcefTerm.AcceptedTaxonID));
      assertEquals("Fabales", tr.get(AcefTerm.Order));
      assertEquals("Fabaceae", tr.get(AcefTerm.Family));
      assertNotNull(tr.get(AcefTerm.Genus));
      assertNotNull(tr.get(AcefTerm.SpeciesEpithet));
      assertNotNull(tr.get(AcefTerm.AuthorString));
      assertNotNull(tr.get(AcefTerm.GSDNameStatus));
      assertEquals("Accepted name", tr.get(AcefTerm.Sp2000NameStatus));
      assertEquals("Terrestrial", tr.get(AcefTerm.LifeZone));
      assertNotNull(tr.get(AcefTerm.SpeciesURL));
      assertNotNull(tr.get(AcefTerm.GSDNameGUID));
      assertEquals("0", tr.get(AcefTerm.IsExtinct));
      assertEquals("0", tr.get(AcefTerm.HasPreHolocene));
      assertEquals("1", tr.get(AcefTerm.HasModern));
    }
    assertEquals(3, counter);
  }

  @Test
  public void corruptFiles() throws Exception {
    AcefReader reader = AcefReader.from(FileUtils.getClasspathFile("acef/corrupt"));

    int counter = 0;
    for (TermRecord tr : reader.read(AcefTerm.AcceptedSpecies)){
      counter++;
      assertNotNull(tr.get(AcefTerm.AcceptedTaxonID));
      assertEquals("Fabales", tr.get(AcefTerm.Order));
      assertEquals("Fabaceae", tr.get(AcefTerm.Family));
      assertNotNull(tr.get(AcefTerm.Genus));
      assertNotNull(tr.get(AcefTerm.SpeciesEpithet));
      assertNotNull(tr.get(AcefTerm.AuthorString));
      assertNotNull(tr.get(AcefTerm.GSDNameStatus));
      assertEquals("Accepted name", tr.get(AcefTerm.Sp2000NameStatus));
      assertEquals("Terrestrial", tr.get(AcefTerm.LifeZone));
      assertNotNull(tr.get(AcefTerm.SpeciesURL));
      assertNotNull(tr.get(AcefTerm.GSDNameGUID));
      assertEquals("0", tr.get(AcefTerm.IsExtinct));
      assertEquals("0", tr.get(AcefTerm.HasPreHolocene));
      assertEquals("1", tr.get(AcefTerm.HasModern));
    }
    assertEquals(3, counter);

    Optional<TermRecord> row = reader.readFirstRow(AcefTerm.CommonNames);
    assertTrue(row.isPresent());

    row = reader.readFirstRow(AcefTerm.Distribution);
    assertFalse(row.isPresent());

    row = reader.readFirstRow(AcefTerm.Reference);
    assertFalse(row.isPresent());

    // try to read bad file
    row = reader.readFirstRow(AcefTerm.SourceDatabase);
    // this somehow works, puzzling but ok ...
    //assertFalse(row.isPresent());

  }

  @Test
  @Ignore("Sth not right how bytes are read from classpath 'files'")
  public void encodings() throws Exception {
    AcefReader reader = AcefReader.from(FileUtils.getClasspathFile("acef/encodings"));

    List<String> expected = Lists.newArrayList("Döring", "(Møllæ) Padléç");
    for (TermRecord tr : reader.read(AcefTerm.AcceptedSpecies)){
      String author = tr.get(AcefTerm.AuthorString);
      System.out.println(author);
      assertTrue(expected.remove(author));
      if (expected.isEmpty()) break;
    }

    Optional<TermRecord> row = reader.readFirstRow(AcefTerm.SourceDatabase);
    assertTrue(row.isPresent());
    assertEquals("Brassicacee species Chöklœst ænd database", row.get().get(AcefTerm.DatabaseFullName));

    expected = Lists.newArrayList("Aparajit", "Gokarni", "Dientón", "Trinchã", "Moingué", "鐮狀狼牙脂鯉");
    for (TermRecord tr : reader.read(AcefTerm.CommonNames)){
      assertTrue(expected.remove(tr.get(AcefTerm.CommonName)));
    }
    assertTrue(expected.isEmpty());
  }

}